backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
pg8000==1.31.2
alembic==1.14.0

# Supabase
supabase==2.10.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Validation
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Testing
pytest==8.3.3
httpx==0.27.2
